Adding outlier detection module and making generic
