+++ libs/datasets/timeseries.py
@@ -932,131 +932,6 @@ def _remove_padded_nans(df, columns):
     return df.reset_index(drop=True)
 
 
-def _calculate_modified_zscore(
-    series: pd.Series, window: int = 10, min_periods=3, ignore_zeros=True
-) -> pd.Series:
-    """Calculates zscore for each point in series comparing current point to past `window` days.
-
-    Each datapoint is compared to the distribution of the past `window` days as long as there are
-    `min_periods` number of non-nan values in the window.
-
-    In the calculation of z-score, zeros are thrown out. This is done to produce better results
-    for regions that regularly report zeros (for instance, RI reports zero new cases on
-    each weekend day).
-
-    Args:
-        series: Series to compute statistics for.
-        window: Size of window to calculate mean and std.
-        min_periods: Number of periods necessary to compute a score - will return nan otherwise.
-        ignore_zeros: If true, zeros are not included in zscore calculation.
-
-    Returns: Array of scores for each datapoint in series.
-    """
-    series = series.copy()
-    if ignore_zeros:
-        series[series == 0] = None
-
-    rolling_series = series.rolling(window=window, min_periods=min_periods)
-    # Shifting one to exclude current datapoint
-    mean = rolling_series.mean().shift(1)
-    std = rolling_series.std(ddof=0).shift(1)
-    z = (series - mean) / std
-    return z.abs()
-
-
-def drop_new_case_outliers(
-    timeseries: MultiRegionDataset, zscore_threshold: float = 8.0, case_threshold: int = 30,
-) -> MultiRegionDataset:
-    """Identifies and drops outliers from the new case series.
-
-    Args:
-        timeseries: Timeseries.
-        zscore_threshold: Z-score threshold.  All new cases with a zscore greater than the
-            threshold will be removed.
-        case_threshold: Min number of cases needed to count as an outlier.
-
-    Returns: timeseries with outliers removed from new_cases.
-    """
-    df_copy = timeseries.timeseries.copy()
-    grouped_df = timeseries.groupby_region()
-
-    zscores = grouped_df[CommonFields.NEW_CASES].apply(_calculate_modified_zscore)
-    to_exclude = (zscores > zscore_threshold) & (df_copy[CommonFields.NEW_CASES] > case_threshold)
-
-    new_tags = taglib.TagCollection()
-    # to_exclude is a Series of bools with the same index as df_copy. Iterate through the index
-    # rows where to_exclude is True.
-    assert to_exclude.index.names == [CommonFields.LOCATION_ID, CommonFields.DATE]
-    for idx, _ in to_exclude[to_exclude].iteritems():
-        new_tags.add(
-            taglib.ZScoreOutlier(
-                date=idx[1], original_observation=df_copy.at[idx, CommonFields.NEW_CASES],
-            ),
-            location_id=idx[0],
-            variable=CommonFields.NEW_CASES,
-        )
-    df_copy.loc[to_exclude, CommonFields.NEW_CASES] = np.nan
-
-    new_timeseries = dataclasses.replace(timeseries, timeseries=df_copy).append_tag_df(
-        new_tags.as_dataframe()
-    )
-
-    return new_timeseries
-
-
-def drop_tail_positivity_outliers(
-    dataset: MultiRegionDataset,
-    zscore_threshold: float = 10.0,
-    diff_threshold_ratio: float = 0.015,
-) -> MultiRegionDataset:
-    """Drops outliers from the test_positivity_7d series, adding tags for removed values.
-
-    Args:
-        dataset:
-        zscore_threshold: Z-score threshold.  All test_positivity_7d values with a zscore greater
-            than the threshold will be removed.
-        diff_threshold_ratio: Minimum difference required for value to be outlier.
-
-    Returns: Dataset with outliers removed from test_positivity_7d.
-    """
-    # TODO(https://trello.com/c/7J2SmDnr): Be more consistent about accessing this data
-    # through wide dates rather than duplicating timeseries.
-    df_copy = dataset.timeseries.copy()
-    grouped_df = dataset.groupby_region()
-
-    def process_series(series):
-        recent_series = series.tail(10)
-        # window of 5 days seems to capture about the right amount of variance.
-        # If window is too large, there may have been a large enough natural shift
-        # in test positivity that recent extreme value looks more noraml.
-        series = _calculate_modified_zscore(recent_series, window=5, ignore_zeros=False)
-        series.index = series.index.get_level_values(CommonFields.DATE)
-        series = series.dropna().last("1D")
-        return series
-
-    zscores = grouped_df[CommonFields.TEST_POSITIVITY_7D].apply(process_series)
-    test_positivity_diffs = df_copy[CommonFields.TEST_POSITIVITY_7D].diff().abs()
-
-    to_exclude = (zscores > zscore_threshold) & (test_positivity_diffs > diff_threshold_ratio)
-
-    new_tags = taglib.TagCollection()
-    # to_exclude is a Series of bools with the same index as df_copy. Iterate through the index
-    # rows where to_exclude is True.
-    assert to_exclude.index.names == [CommonFields.LOCATION_ID, CommonFields.DATE]
-    for idx, _ in to_exclude[to_exclude].iteritems():
-        new_tags.add(
-            taglib.ZScoreOutlier(
-                date=idx[1], original_observation=df_copy.at[idx, CommonFields.TEST_POSITIVITY_7D],
-            ),
-            location_id=idx[0],
-            variable=CommonFields.TEST_POSITIVITY_7D,
-        )
-
-    df_copy.loc[to_exclude[to_exclude].index, CommonFields.TEST_POSITIVITY_7D] = np.nan
-
-    return dataclasses.replace(dataset, timeseries=df_copy).append_tag_df(new_tags.as_dataframe())
-
-
 def drop_regions_without_population(
     mrts: MultiRegionDataset,
     known_location_id_to_drop: Sequence[str],
